Reduce use of exported_system_prop
Reduce use of "exported_system_prop" by defining a new (currently
identical) "system_setting_prop" type for the system properties that are
for "global system settings". See the comments in
private/property_contexts for details.

Initially the rights of the new type should be identical to
exported_system_prop but they will be reduced with a follow-up commit to
enable easier rollback / progress to be made on related work.

Bug: 236612872
Test: treehugger
Change-Id: I8d818342023bc462376c091b8a522532ccaf15d3