Revert "Limit processes that can change settings sysprops"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/a594ecf5-c964-471b-aff6-082fe91770a8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4ddc7ed9a28e70e7a4211bbd5035abfd8c19a4ac