Revert "Add new type for system settings metadata"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/a594ecf5-c964-471b-aff6-082fe91770a8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I45bdc38beddbd28b6d3f578ce7ee3661939b63cf