Grant appdomain delete_name permission on user_profile_data_file.
To prevent race condition on a profile, the app holds a flock when writing the
profile, and profman needs to hold a flock to read it. This is not ideal because
either side can get blocked by the flock.

We want to avoid using flock and do it in a move-based way: instead of mutating
the profile in place, the app creates a temp file next to it, works on the temp
file, and replaces the original file after it's done (or deletes the temp file
if it fails).

To achieve that, the app needs the delete_name permission.

Bug: 249522285
Change-Id: I16f27e6a9c5c3a7ab2ab8e24d3ad0a20119e16db
Test: Presubmit