Revert "Move DICE logic into microdroid_manager"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/f1d4b608-68bc-4be9-aa6d-61dc5728b43f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibd02d935c33ac0af05d3723f0e4ce38296e830dd