Add rules for cgroupof binary
The binary is to read /proc/<pid>/cgroup. It is made as a binary in
order not to require the caller to have privileges to read the files.
Specifically, the caller needs to 1) be in the AID_READPROC group to not
affected by the hidepid=2 option of the proc mount and 2) have read
access to domain:file and domain:dir because /proc/<pid>/cgroup files
are labeled as the type of process they are for.

This CL defines cgroupof_exec and cgroupof types and give it minimal
privilege to read the proc entries. It then allows virtualizationservice
the privilege to execute the binary and run it in cgroupof domain.
Finally, it adds an file-contexts entry for the binary in the virt APEX.

Bug: 245727626
Test: run a VM. The spanwed crosvm process is in the correct cgroup
Change-Id: Ib47024d1ae30db86629e256c6235fcfa02c0f30c