Create virtmgr domain and initial policy
Start a new security domain for virtmgr - a child proces of an app that
manages its virtual machines. To get started, populate its policy with
a clone of virtualizationservice's. As we go forwards, we'll remove
permissions from those domains depending on their responsibilities.

Also add permissions to auto-transition to the virtmgr domain when the
client fork/execs virtmgr, to communicate over UDS, and parent->child
SIGTERM for termination and parent-death notification.

Bug: 250685929
Test: atest -p packages/modules/Virtualization:avf-presubmit
Change-Id: I7624700b263f49264812e9bca6b83a003cc929be