Revert "microdroid: Allow microdroid_manager to get local CID"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b86191be-4f6c-4874-99ab-94062fc117f7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic57d8bfa9b6e6212403e19b122d7b1799071aa37