Allow system_server to read /data/misc/profman.
Before this change, system_server only has write access. We want read
access the directory so that we can check if it has the right
permissions before we write to it. This change also grants system_server
read access to the files in that directory, which is not needed now but
gives us the flexibility to read the files in future ART versions, which
can be pushed by mainline updates.

Bug: 262230400
Test: No longer see SELinux denials on that directory.
Change-Id: Ic26b2a170031c4f14423b8b1f1a8564d64f532ae