Revert "Add usbd servicemanager permission"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a48f3bcf-0f95-40d6-8117-3d0ed5fc50aa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id552d480003336ea4d0567abbd0e07d6a13da973