Revert "Add SEPolicy tags for concrypt cacerts."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4a44bc4c-8cf9-40c0-a662-3c58d8e4d314). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I946466e1922a5c7ba4519e05e0d3076abc39b62c