Revert "Add SEPolicy tags for concrypt cacerts."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2b9153a1-1364-446b-9f54-9eee1eb68233). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6e1efc39e12cbdeb6a548b533e001b56b66a517b