Revert "Add missing permissions for default bluetooth hal"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b68dd3fa-1238-4b18-b603-65cf134ca943). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d87c5330c18337f67828334571d74e0caf09e36