Revert "Add tombstone_transmit init property to microdroid"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b7721fd-2720-4362-bf72-0794e238fcc1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5748131c6446d5a73de47b1a03fe10a01eeaf80e