Revert "Update sepolicy prebuilts for PRNG seeder changes."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89dd3ee0-9d24-491b-a876-43fc01c409b7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib6f86437748c142d0f9867cd9f624a1ac05ebd54