Revert "[MTE] Add memory_safety_native_boot namespace"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5653ec87-ac06-4d7e-b959-cf0c919ef321). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If074b022b2b5e43ecf1587b15fe493c85017a98b