Revert "Share isolated properties across islolated apps"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8e6c6976-cbb0-4d81-9d5c-68546037a54e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibe7b02250425a64c34e91fbf432cbb6c3f267313