Revert "Add SEPolicy for PRNG seeder daemon."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/89a9c71a-8217-452c-b41a-2158b1420baa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I094c1996e7859e5af3cc62bd7b8162485dafc4f6