Revert "Add SEPolicy for PRNG seeder daemon."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b89bd1b7-a3df-45e9-abd6-259e51179920). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia44f632fe57bf71dfa8f07b954c7809e5c900d4d