Revert "Allow virtualizationmanager to read AVF debug policy"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/781f46d4-e6d9-4c5b-abf8-c61e6a026419). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia4027dc40f149f35d759485125153b9a7e009f7f