Revert "Add bluetooth_prop to system_server sepolicy."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aa5aa3a2-3d8f-4a40-b864-c7c1a68417fe). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib0172ca49a1722b472b6aad442f2d65544f20ef0