Revert "Fix incorrect domain used in system_server.te"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4eb3030f-f87e-4362-b4e9-91efb34fa22b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7854f1d6119eb3a01b8907821d6712b358367c6f