Revert "Remove implicit access for isolated_app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5d486c6f-9057-46cb-9793-1d908470d224). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id3f745e26759cbc6cb0156479f42f463d212a0a0