Revert "Remove implicit access for isolated_app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/44597106-7d76-43b5-8901-49b92beb8c20). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I02bfe17373ea9a84e3085171b41209e107d19c83