Revert "Remove implicit access for isolated_app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0f2fe2fe-d017-4845-a3f4-1cf581d19da0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie1850cf618628cc91bdec676de6e88b795ed9ccc