Revert "Remove implicit access for isolated_app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9969a5b5-ff6d-4aa4-bf79-cf1cb8153ca2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8e0e74edda0333046d06612fb3acfb487a1e4fcb