Revert "Remove implicit access for isolated_app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/05b34ad6-7f87-4467-8fc6-6e4e893545d2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0541bcd7acbc9cf8a5780ef8ab4c0a226531064f