Revert "[dice] Remove all the sepolicy relating the hal service dice"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43a314ea-5852-4e51-9bae-47079a9ff8d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I309951346a42f854bd24fe4dd037235f4ea2b9fe