Revert "Remove 34.0 compat modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7bf1f393-9e63-429b-8598-3772362f9e01). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If388b48525fc72fa9e58746697deec0194a4d4fc