Revert "Remove 34.0 compat modules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6340655f-10ac-4e8e-8f38-240cd57d559e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3fcdfbad156004b41d7c4575c61461f857caebd4