Revert "Remove implicit access for isolated_app"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cf0c51ae-7e34-4633-8049-17797d5c4230). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If0069e6496a9b5b807d66005b411b7992c0545b6