Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2ea5aceb-9086-491a-a0d5-4a612c6697ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I44bda6fa0c7391b2b2fb4899ff3e29fd1ba95536