Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/772a2377-e18a-45f5-86c1-fb3e7d3f0d88). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic37df78f87dad95b9d80b2a91ba17dafc9a983fc