Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/772a2377-e18a-45f5-86c1-fb3e7d3f0d88). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I61e52057a0d056111c0143466f9e195b48e051d9