Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b2b34415-ab3d-4ae6-ae1c-b99829b120bd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib47163e4cd30814eab9732e542b0fa6b092b4d0b