Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/42be7ae3-e8ac-4843-b96e-5349ab99a53e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia6cec8755f035302535bb5e3e0dc61b946e8e344