Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd3acea4-8197-4c99-a651-5b7bd458228f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I51216a48ac0c98eda2b8a5a31eb0280e3fbb00cc