Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/782b5332-3a36-4951-9e29-08e1aa941fb5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I85847828e7c057323cce8d4f7eacd0a466ec36ba