Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5575dad3-1a97-4522-8a9d-c1fbec8b2bf8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia9cd21cac6997e873d6d813f9a81874fbce6d357