Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4eb21153-673a-44c8-a8d3-cc0a8a88622f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1c2a2c6e6f985c2ce77be4817070b2ed3b2ceca3