Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e2b33ed6-4f34-4c7d-8570-e48c966ed80c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I90412077cc6576e5f918743e11093230df9f1f15