Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/65811683-f9ed-42ee-b33f-aeb0871af156). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I62ccafabe2af5c6fa542889a5a9f0ead2cd95976