Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3d391edc-7e68-4c1b-889d-81885189c5d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibbc22ebb1a4d0b4c6e51fe10c292447a3f08a9f9