Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/95b76727-8531-4878-bb76-77260be1dbe0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idbcb67885bac7a78ad2048aeeddda2ade5a64a76