Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f4250ad-8977-4607-90ab-dbb69bef95ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I900ab955bcf2929a0d48eb6fc3230dfcd0dd8f0f