Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/80eee2d8-b873-477f-9cfa-4ca6896faf93). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ca200eaa431dd01d22a8da1e2abd9cd444cd4a7