Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb86f2c4-200f-4908-96c1-d67368b6bf7a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0906d3bcfa01c3e3beffac3326f3c24465cf818a