Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/28ded5ef-7063-4213-9390-4e2412ce6801). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I178d8c3727ff0294c6bab2c495f10e0b4deb32f4