Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/78e2c66c-1b84-4de0-a0fa-e81d76785b8e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iba336b184104fdf842e8f8bef18de4eb1a38853c