Revert "Introduce a new sdk_sandbox domain"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/33d9a64b-944c-4f79-89d6-939836efec80). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifaf605b472e51a4868dacce3281ab8067bcd7eb7