Revert "Create sdk_sandbox_all."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/14a163f0-efc2-497c-8806-8fc197d9230d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie84522752654ffa9fbc56f6bcf1e1b2ef83bca05