Revert "Create sdk_sandbox_all."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/acdf3fcd-a54c-4f49-957f-6ad441baa3f8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2887af8884cd4b8430cdcea3e323293ddcf538b1