Revert "Parallelize singleton execution."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6c7fd9b5-a956-4458-afaf-8aca15a24eb8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id804523bb348b65be39785d7faccee561da198ba