Revert "Parallelize singleton execution."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9d15aa48-4d40-4853-8c06-68fc4c967f90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6915b25522c345d34ce3330c55c7068681b2af52