Add SELinux policy for storage areas
We are adding the ability for apps to create "storage areas", which are
transparently encrypted directories protected by screen lock bound keys.
This CL makes the required SELinux policy changes.

First, assign the type "system_userdir_file" to the new top-level
directory /data/storage_area (non-recursively).  This is the same type
used by the other top-level directories containing app data, such as
/data/user, and it restricts access to the directory in the desired way.
Note that it gets the categories of the user.

Bug: 325121608
Test: atest StorageAreaTest
Ignore-AOSP-First: other changes in topic have conflicts with AOSP
Change-Id: I74805d249f59226fc6963693f682c70949bfad93