Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b5677ee6-2d90-4b60-8e43-a0195b95d09a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I10361b85b302d7fbb3904a675bdc812d3a262fa8