Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/60683d0d-dd28-4043-9758-cfc98ea00cd6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I636b15c924e9024792a54a3aa124f25d31728a9a