Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a0155bb6-b15a-455b-b3f3-ac4c2a6a18e9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0d36b084155c3cc97e08cab5a8ebcd2315666542