Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/831b6814-17d2-4fb0-a920-f0119fea3b78). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice345ca08df62abd6cc6cbe3f2b55b3d3cf769b4