Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d94d720-79ef-4634-bcdc-211452a65f57). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic3b523c151993fc95a29b7b114d6ed9d5c250444