Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ea4cd69a-ccd6-4824-8764-f988c47f5b56). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0ac0ae5ca647465ca6085f8029b36ef7bd894eec