Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/124e6d69-6d95-4014-a867-d6f3e606feed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I18266b7ef34b99f0d3c95f78e2ea7d769402a182