Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/31c5cef5-db56-4e93-82d7-e8194036bee4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia782bd38f3e76bf08e21a2e70050911f0deeea95