Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2c93eb11-6dc1-4057-b3c7-f386a1d16f1e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icee1a26cf3e2e987b27ab6a48b35e14dd6168751