Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c768d6c0-3c8b-4bd9-a052-41c7c8bc62b3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I991c0142ac767f09622e7fef4734e3867552c115