Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9289a4e6-02e0-44f9-9540-c1832ed19129). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I99a7092c285ff8c6594ab5c5b4a66fe3c87c187f