Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/013ad17f-2683-4faa-9ef0-7e09a9e1ddad). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I49bb5539ce664f16e7114b706db1a84c6a473f85