Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a2606e11-4cd0-4438-b7b4-a70e239a4bf2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id6608e3711d1331ae76a616fed8da0f8a5ccd66d