Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3386f164-1cab-434f-a69e-70c9a90d5307). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I183c30b8aaa81307004bf772e803ad19af07af4b