Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3337dd48-4956-415b-b9b4-c8ec6935b08c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0127b674f583f4a2ce5508e186c3c511ebd0a1c3