Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/dcc14859-d833-4842-8d75-246f9b85dfcb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I019da2545a51c93a247fcc6559b9e74c83ee21a6