Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe733ec3-c45e-4252-98f0-2b284b6f67a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib210b59d6dcbee780fc9c6e52363caa36cc868e7