Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/86aad5a6-9d26-4efe-8f42-bfb0f3e27b44). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I40933b72f92477b84852eaf50e4d224624973bc6