Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e76d9a7f-612f-46dc-ab76-66c7e89c2983). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia86279d8b62ea87701d22a7081eee0a9dbecc79f