Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/07ba2557-84b6-4b59-b96c-a3e01aa339e6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7e3626c48d0353d1aee5a9a41ee848c3bd2a799a