Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a80a4fae-fd59-4fdf-9074-1183ec61cde6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I21095c605df76db900e2e9373c49f9a670249fd4