Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4b07ea00-3186-4841-b454-ea3fcde6e765). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idc52eff41e0cfd5a6ea1a08f0247be02b70e1326