Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0efe3abd-bdd0-49e1-b022-e6b2b4a14e81). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7c5eaf1c40c711f40866ef5fa89aa1abe91bf508