Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18f8ae8c-cd8a-4679-af75-74b17c94aaae). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib845184018a96b2c3775db2292a135d0b51440c0