Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/241f09b2-d46b-4afc-965d-32e976ef5b14). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I69d92ce6994bd30d7f8d5312e19afa022591c956