Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2ec2edf1-5a29-4ccc-9153-84b0552703b5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I18724c8ef5b37eda1531c6079a297eece33ead50