Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bb90b4da-f71d-4777-91c4-f474147f26af). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia500d218ec9f8c5afdf6709d3385912662f9cf94