Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b9bca9ba-c125-4d2e-803d-9bfb32b8a76b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id0d758a1b9278d5b083ecd033d21dc6be5c1f481