Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/75f078e3-904f-47d7-b93e-9e88aeab332c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1cdbe705372e7e4a50f55651eae6d2f12c6fcf29