Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ec943aca-601d-4ad9-a291-bfa188ca0789). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I637babdcbc49e4de41508fd01ba0f017a0b85f36