Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cd70e8b2-ad9b-4b80-8adc-449e854585eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6325b8121263f40e49dfc6e0ad97af7da28d9219