Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9f8d9df7-6003-45a7-8866-c42dd5208c71). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I904e9414d5c83d0bc1b872a38251629dcfd55096