Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/44622b5e-9d5f-4bcc-b657-5df275bba5f3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibc1a0f74746158c77cde19d39bbdd4a468bb74b5