Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/04575a1d-c1df-4727-b2ea-0fc6f75db5b3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4484f03e2e8feb3e167a3d5c1b0a2023cc0f9ead