Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81033931-ad35-4a62-9695-74fd5c8cb29d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3e909dcf26499bfb40e6d1830bdd5a9a0f63900c