Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/84d3ddb3-a437-415c-919c-74c990b1568d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa4efb351ae6f56d00e633ee5d66eb2ed1eed986