Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f763770f-48ec-42a5-a9c7-dbdcbd582c90). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I36f140eb3d7e4983b6c1ada4568897d3ed9e958b