Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8dd65dae-4d2d-432b-ad8b-1ed493b312eb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I11738953cb12e182aa305be44b98fa4942f736d8