Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b00d6121-a3c8-44b9-bdcc-456d9b92f2ba). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icea5062574c4943d2a583bdb9b9ce4d9bf266896