Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e37e88a7-0065-4fb5-8bbd-9724260e1c1f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If80da272b16551024da3795ee30c060bd1c91160