Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c0ee7779-9e8f-4dbc-b332-bbad55be73b6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice51e5d14f129a9137052c54943e3346009901c6