Revert "Introduce vendor_apex_metadata_file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2e268ef4-d6e2-4dda-bd7a-838bf8a4bf82). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I29c1e7247d6c6dfdf9cbeb20d7fea38a84576b36