Revert "persist.sysui.notification.ranking_update_ashmem"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7df96236-afe6-48d3-84e7-d7c5304a8c1b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22aa455275824a2294bc137589dadf6c89085287