Revert "Fix apex_sepolicy_tests_test"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/566474e9-edf0-4104-b271-dba6cbe19300). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia1a5d957447c7aaa2b300aa28d6b8c6837b48006