SELinux: Give vold additional permissions
..on media_userdir_file file type. Currently, the migrate
storage operation is failing due to lack of permissions
on media_userdir_file.

During migrate data operation vold performs a copy (using cp) and
deletes the copied directories and files from the previous volume
(or the source volume). A new file type was defined from CE
directories in CL [1] for files or directories which resides in
primary storage (or /data/media or /mnt/expand/<UUID>/media), but
appropiate permissions for cp command was not given to operate
on those files. The same applies for the rmdir permission which is
required for cleanup during the migrate operation.
This change adds the permissions rmdir permission for directories in
media_userdir_file file type and create_file_perms for files for the
same file type.

[1]: I1fbdd180f56dd2fe4703763936f5850cef8ab0ba

Test: manual
Checked that on executing "Move Contents" operation from Settings
there is no avc denied logs and the operation is sucessful.

Change-Id: Iabcddd20869e5f7e17cef850344cbef8dd6eef34