Re-land "Remove fsverity_init SELinux rules"
Since the fsverity_init binary is being removed, remove the
corresponding SELinux rules too.

For now, keep the rule "allow domain kernel:key search", which existed
to allow the fsverity keyring to be searched.  It turns out to actually
be needed for a bit more than that.  We should be able to replace it
with something more precise, but we need to be careful.

[The topic containing the original version of this change was reverted
because it had a prerequisite that hadn't reached AOSP yet.  That
prerequisite has now landed, so the topic can be re-landed now.]

Bug: 290064770
Test: Verified no SELinux denials when booting Cuttlefish
Change-Id: I27fc3d0b4018e877a5d3c78d3d95de4ddbc145fb