Remove fsverity_init SELinux rules
Since the fsverity_init binary is being removed, the corresponding
SELinux rules can be removed too.

Also remove the SELinux rules that granted all domains permission to
search the ".fs-verity" keyring.  This keyring is now always empty or
nonexistent.  And it should never be searched at all, since files with
an fsverity builtin signature aren't expected to still exist on-disk.

Bug: 290064770
Test: presubmit
Change-Id: I992b75808284cb8a3c26a84be548390193113668