Hide font_fallback.xml from applications
font_fallback.xml is a new system font configuration file as
a replacement of fonts.xml.

Historically, the fonts.xml has been directly read from the
random applications and it has been a blocker of adding new
feature to the fonts.xml for the compatibility reasons.

To encourage using public APIs for getting system font
configuration, making this new configuration file unreadable
to the applications.

It is necessary to be read it by zygote for preloading some font
files for the performance reasons.

Bug: 281769620
Test: atest CtsGraphicsTestCases
Test: atest CtsTextTestCases
Change-Id: I05011c9313fa3818ec50d9884227512ef1b0fda9