Revert "Allow lpdumpd to read Virtual A/B diagnostics."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6d1f425d-0ff2-4be6-bdfa-5bce548cb790). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic5d56783f2d91532b0d537690f7cd3b04fb67327