Revert "Remove fsverity_init SELinux rules"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4bc61e87-8d06-4c8f-9475-defadf2bc42f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I72d393c16828f7570c591083960f4670e15efe84