Revert "Add /bootstrap-apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4d84082-d35f-4b2b-8f9c-61159010f979). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I248996fb517611b213a3e78075f3e7a044365565