Revert "Add /bootstrap-apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ff01a85f-3173-4938-b77d-ac68725b2869). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iee6d67a00cf73d4f5c34c8e74fc7bfe09d9a0cb0