Revert "Add /bootstrap-apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aec73b48-0156-4581-b6cb-eb5b80422685). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I551bec6fdc6d07cfb478a7ab9f1aa3ca36301b8b