Use apex_service to get the list of apexes
vold runs in the bootstrap mount namespace. Since aosp/2666915,
vold(and its children) can't see the other apexes activated in the
default mount namespace. This caused vold_prepare_subdirs to fail to
create apexdata directories for those non-bootstrap apexes. It's because
vold_prepare_subdirs relis on scanning /apex directory to get the list
of apexes. It was accidentally okay because even in the bootstrap mount
namespace directories for non-bootstrap apexes were visible (as empty
directories, not correctly mounted).

In this change, the logic to get the list of apexes are replaced with
the query to apex_service, which is the source of truth. This also helps
to remove the manual handling of the internals about how apexes are
activated. (Note how it filtered out directories with "@" char).

Bug: 293949266
Bug: 293546778
Test: CtsPackageSettingHostTestCases
Change-Id: I86f58158b97463206fb76a0c31f29b78874f4c35