Revert "Add /bootstrap-apex"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4eaa9635-f4d2-4ce4-acdf-19f726beae52). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie451c624442189ccaf5b94d924df76aec1ecb5da